+++ UnityProject/Assets/GameScripts/HotFix/GameLogic/DataCenter/PlayerData/BasePlayerData.cs
@@ -1,0 +1,13 @@
+namespace GameLogic
+{
+    /// <summary>
+    /// 玩家数据基类。
+    /// </summary>
+    public class BasePlayerData
+    {
+        /// <summary>
+        /// 角色ID。
+        /// </summary>
+        protected ulong m_roleID;
+    }
+}